refactor(backend): extract code studio scaffold shell
- move shared scaffold palette, title, aria, CSS, script wrapper, and shell composition helpers into code_studio_scaffold_shell.py
- preserve Code Studio scaffold public API and renderer wrapper names
- update cleanup audit and codebase navigation docs
